:recycle: [open-formulieren/open-forms#4929] Move data-layer interaction into separate module
This makes it a bit easier to manage the pure React stuff in
the component file before we actually replace the props with necessary
context.
